+++ front/src/Apps/common/Components/CompanySelector/CompanySelector.tsx
@@ -13,12 +13,13 @@ const isSearchValid = searchClue => searchClue.length >= 3;
 
 const CompanySelector = ({
   loading,
-  onSelect,
-  onSearch,
+  selectedCompany,
   companies,
   favorites,
   disabled = false,
-  selectedCompany
+  searchLabel,
+  onSelect,
+  onSearch
 }: CompanySelectorProps) => {
   const [searchString, setSearchString] = useState("");
   const [postalCodeString, setPostalCodeString] = useState("");
@@ -95,9 +96,9 @@ const CompanySelector = ({
       className="fr-container--fluid company-selector"
     >
       <div className="fr-grid-row fr-grid-row--gutters fr-grid-row--bottom company-selector-search">
-        <div className="fr-col-8">
+        <div className="fr-col-12 fr-col-md-8">
           <Input
-            label="Numéro de SIRET ou nom de l'entreprise"
+            label={searchLabel}
             disabled={disabled}
             nativeInputProps={{
               value: searchString,
@@ -113,7 +114,7 @@ const CompanySelector = ({
             }}
           />
         </div>
-        <div className="fr-col-4">
+        <div className="fr-col-12 fr-col-md-4">
           <Input
             label="Département ou code postal"
             disabled={disabled}